requirements.txt
alembic==1.11.2
aiohttp==3.8.5
arq-prometheus==0.3.0
aiosignal==1.3.1
aio-pika==9.4.0
asyncpg==0.29.0
asgi-lifespan==2.1.0
annotated-types==0.5.0
async-timeout==4.0.3
bcrypt==4.0.1
fastapi==0.110.0
frozenlist==1.4.0
greenlet==2.0.2
multidict==6.0.4
loguru==0.7.1
passlib==1.7.4
premailer==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
py-grpc-prometheus==0.7.0
pyaes==1.6.1
pyasn1==0.5.0
pydantic===2.3.0
pydantic-settings===2.0.2
pydantic_core==2.6.3
pytest==7.4.4
psycopg2-binary==2.9.9
pytest_asyncio==0.23.3
python-dateutil==2.8.2
python-dotenv==1.0.0
prometheus-fastapi-instrumentator==6.1.0
python-jose==3.3.0
pytz==2023.3
httpx==0.27.0
starlette==0.36.3
SQLAlchemy==2.0.19
uvicorn==0.23.1
vine==5.0.0
mock==5.1.0
wcwidth==0.2.6
yarl==1.9.2
arq==0.25.0